MachineLearning/ArchivosParaConstruirImagenDockerCovid/requirements.txt
# Archivo 2
fastapi
uvicorn
numpy
pillow
tensorflow
python-multipart

# pip install -r .\requirements.txt

